IControlAdvancedFeatures: replace error check with doc comment
Checking that all instances of IControlAdvancedFeatures are root nodes
is not helpful because some of them may be used in tests, generators,
etc.

In general, this warning is appropriate for language designers, not
language users, so I'll assume language designers will look at the
concept declaration and see the documentation.