expr.temporal: make use of temporal context
IHaveOptionalPartsInTemporalContext can be used to maked children of a
concept as required if there is no temporal context proviced by an
ancestor. A node factory will take care of creating these chlilden in
contexts where they are required so the user guided to enter them and
is not confronted with a typesystem error instantly.