core.expr: add custom error message to constrains
Added a new child to the ConstrainItem to allow the user to specify
a optional custom error message. The interpreter will interpret the
expression placed there by the user to get the custom error message
and display it when the constrain was violated. If the user did not
specidy and custom error message the same behaviour as before will take
place and the user is presented generic cosntrain violated error.