comp: introduce base concept for all kind of connectors
In order to allow a more abstract representation of connectors for a new
connector `AbstractConnectorBase` was introduced. This concept assumes
that at least a source port (governing port) exists. The target for this
connector is still a abstract node. It is usefull in cases where a the
target of the connection is a bus or some other target that is not a port.

In additon the `AbstractConnector` was renamed to `AbstractPortToPortConnector`
It is a specialisation of the `AbstractConnectorBase` which asumes a that
the connection is made between two ports. Most of the behaviour methods
and interface implementations have been moved up to `AbstractConnectorBase`.

The `AbstractConnectorRefTarget` was changed to point to a `AbstractConnectorBase`.

Also the typesystem rules from `AbstractConnector` are moved up to the
`AbstractConnectorBase` except for the checking rule that checks the port
to port connection.