comp: add parameter dependencies for ports
Includes a new port type that can wrap a IPortType. It can reference a
parameter. When using these ports in connectors the type system ensures
that only ports with same payload and configuration can be connected.

I removed the for "is statically evaluatable" from the checking rule of
a component instance. Since that property of an expression in not
maintained at all it will produce a lot of false prositives.