comp: add better error message for invalids ports
Moved the typesystem rule form assembly connector to the
AbstractPortToPort connector. This allows us to check all connectors
genericly.

Better error messages for cases where the configurations datatypse used
for the ports in the connection don't match. The helper class tries to
extract common usecases, for now just RecordsTypes, but could get
extended in the future. Right now there no way to customize the error
msg in a language extension.