core.base: move IControlAdvancedFeatures to core.base
Moved IControlAdvancedFeatures into core.base. In order to allow this it
was required to change it's constraints implementation. The dependnecies
needed to be inverted because the constrains otherwise would depend on
cocepts from the expression langauge. The constrains were moved to the
individual concepts and implemented as a can be child constrain that
checks if the IControlAdvancedFeatures root permits the use of the
category that the feature belongs to (e.g. attempt types). As a general
pattern we should stick to only declaring the categories on the interface
and then check for it in  concrete cases. That way we prevent a dependency
on these concepts.