build: easier branch handling for mbeddr
At the moment building against specific branches from mbeddr is a very
tedious task. The user has to "escape" the branch name and take care of
adding the trailing ".". This has been simplifies the user can simply
assign the branch name to the new "mbeddrBranch" variable and the script
will take care of the rest.

Additionally some more logging was added to identify if certain
properties have been overriden by the caller of the build script to
make debugging easier.