components.core: removed IComponentsContent checking rule
If one inplements the dependencies() behaviour method,
this would imply that if you chunk content is public all
of its dependencies should be public at all.

I think this depends quite a lot on the particular implementation
of the chunk content and therefore I remove this checking rule to
at least give a change to the particular implementation to decide
how to handle these cases.