expr.data: allow lookups in data tables
New lookUp operation allowed on data tables. The operation
will look up the row based on the column and value provided
to the expression.

New flag "allows lookup" on data tables which is at the
moment only a flag to allow the operation. If the flag is
present the type system provides an info message about the
uniquness requirements for values in a column. Right now no
static checks for uniquness of all values in a column exists.
Since each cell can contain an arbiatary expression we would
need to run the interpreter in the checking rule to find
duplicagte values.