expr.base: collect coverage in each DefaultCoverageAnalyzer instance
In order to be able to collect coverage information for a single execution of the interpreter and to compare
different execution of the generator it is required that the information about which nodes where visited is
held in each instance of the DefaultCoverageAnalyzer. The DefaultCoverageAnalyzer now also holds the indivdual
values that node was evaluated to. Right now this is quite some duplication with the ComputationTrace which only
hold the values but no other coverage information. Both should be unified into a single data structure in the
future. The major concern why this wasn't done as part of these efforts is that the data structures are
optimized for different use cases. The ComputationTrace is easy to navigate along the tree structure of the
model while the CoverageInformation are optimized for doing random look ups for nodes and their values.

The static API for "global" coverage information was kept usable. These will access the information stored
in a static instance. This allows the editor coloring to work as before.