expr: coputation trace extensions for dectabs
Introduce additional traces for decision tables similart to MultiDecTab in the computation
trace to allow easier analyses of the matched row & column in coverage analyses which use
the computation trace to calculate the coverage.

DefaulCoverageAnalyser now allows to store more than one value for a node.

By default only the last value is returned but a new instance method to retrive all values
was added.