Units: make units migration rerunable
A bug in the units migration didn't set the now deprecated "description" property to null therefor
every execution of the rerunable migration would create new quantities. The migration was updated
to detect this case properly and not change the model in this case. In addition the migration now
sets the deprecated property to null.

fixes #423