kernelF: fix records equals with temporals + extension point
Adding the temporal lanugage to a model changed the behavior of the
BinaryEqualityExpression behaviour. The TS is fixed, so that the
behavior is now independent of the temporal language being present.

For The Records isComparable() method, an extension point is introduced,
so the behavior of the type system can be changed in customer projects
at that point