core.base: add IEmptyLiteral
IEmptyLiteral is a marker interface used to mark literals that at runtime
evaluate to "nix" in order to use this information in static checking.
The marker interface allows static checks in kernelf while allowing languages
that extend kernelf to contribute their own domainspecific keywords for empty
and still inherit the checkings.

The first subsystem to use the checking is expr.data. There are checking rules
to detect common pitfalls in data tables like multiple empty values in the same row.

Implementation of the interpreter fixes a left over from the nix migration
where the data tabel interpreter would incorrectly return null
instead of the correct runtime representation causing runtime errors in the
interpreter.