expr.data: add emptyInput literal
The `emptyInput` literal is semantically a shortcut for `emtpty<type>`
the can only be used in data tables. At runtime it is represented as None/Nix.
Implementation of the interpreter also fixes a left over from the none/nix
migration where the data tabel interpreter would incorrectly return null
instead of the correct runtime representation causing runtime errors in the
interpreter.

The literal comes with checking rules to detect common pitfalls like multiple
emptyinputs in the same row.

The rational behind introducing a specific literal is mostly lacking infrastruture
to allow `empty´ without type annotations and the requirement in upstream projects
to detect the "default value" when data tables are imported from external datasources.