core.base: add IEmptyLiteral
IEmptyLiteral is a marker interface used to mark literals that at runtime
evaluate to "none/nix" in order to use this information in static checking.
The marker interface allows static checks in kernelf while allowing languages
that extend kernelf to contribute their own domainspecific keywords for empty
and still inherit the checkings.

The first subsystem to use the checking is expr.data.