expr.test: improve scoping performance
Replaced `nodesIncludingImported` and `where( x.InstanceOf)` combination
with a direct call to `SModelOperations.nodesIncludingImported` in the
TestSuite scope implementation. This change allows MPS to cache the model
and concept tuple and improves scoping performance significant. In one
project it reduced the model checking time by 30%.

MPS doesn't support passing an expression to the concept part of
nodesIncludingImported but only allows direct references to concepts and
calling the underlying API directly was required.