[breaking] Extract shared runtime from base runtime
Shared runtime contains classes that will be used both from the
generated code and by the interpreter or in-MPS code.

Even though the individual classes used by the generated code do not
depend on MPS-specific classes, extracting them from the base runtime
means that the generated code no longer needs to depend on a module that
contain MPS-specific code (e.g. NixSupport with its usage of the node<>
type).

Breaking change: class INixValue was moved from the existing solution to
the new one so references to it will need to be updated (Ctrl+R to
import org.iets3.core.expr.base.shared.runtime model, then F5 to update
the reference).