Remove MPS dependency from the published POMs, add tests -> opensource dependency
While it's a good idea to know which version of MPS a given version
uses, this is currently solved by the version number. Having MPS in
dependencies causes problems when downstream projects want to use Maven
to unpack all their dependencies to use as project libraries in MPS.

A version of MPS would get unpacked among other dependencies such as
mbeddr, and the dependency on MPS would have to be manually excluded to
prevent this.

On the other hand, the tests artifact was missing the dependency on
the opensource artifact, this commit adds it.