Deleted all empty generators
When a custom generation plan is defined the model checker shows the error
'model uses language X with 1 generator(s), none of which is engaged with the active generation plan'.

A workaround would be to just include these languages in the generation
plan, but since they don't do anything it's unnecessary noise.