datetime: add concepts for time
Added time support similar to date support. Time is time on the "wall clock"
wihtout a timezone. There is no support for "datetime" which would be a
date with an associated time.

The usual operations are implemented: ´.hours`, `.minutes`and `.seconds`.
Time literals start with a backslash (\) rather than a forward one for date.

Literals for hours, minutes and seconds are supported and are usable to perform
arithmetic operations with times like `\22:00:00\ - 2 hours`.

There is no generator support at the moment.