datetime: add generator for time concepts
New java generator for time concepts. The structure, especially the order in
template switches tries to closely resemble the structure of the existing
generator.
In some cases that might seem odd especially for the DotExpression generator
because for instances a `HoursValue` target could never be part of a dot that
doesn't have a TimeType on the left so the check isn't required. But for
consistency the same pattern as in the date generator is used.