InterpreterRecordrenderer: fix links to model for (Short)LambdaExpression traces
Problem: Though the interpreter shows the traces for (Short)LambdaExpression-calculations for each parameter-value,
these traces were not correctly linked, so it was not possible to go to the associated modelpart from the trace.
The interpreter works on copies, which are not part of the model, of (Short)LambdaExpressions and linked the traces to these copies.

Solution: Link the traces to the original Expression of the copies.